samples fully debugged
